+++ splashsurf_lib/tests/integration_tests/mod.rs
@@ -1,5 +1,5 @@
+#[cfg(feature = "vtk_extras")]
 pub mod io;
-
 #[cfg(feature = "vtk_extras")]
 pub mod test_full;
 #[cfg(feature = "vtk_extras")]
